requirements.txt
pymupdf
langchain
openai
chromadb
pyyaml
python-dotenv
langchain-community
tiktoken
langchain-chroma
langchain-openai